refactoring: moved lib python files in submodules
