+++ exasol_script_languages_developer_sandbox/cli/commands/setup_ec2.py
@@ -7,8 +7,8 @@ from exasol_script_languages_developer_sandbox.cli.options.ec2_options import ec
 from exasol_script_languages_developer_sandbox.cli.options.id_options import id_options
 from exasol_script_languages_developer_sandbox.cli.options.logging import logging_options, set_log_level
 from exasol_script_languages_developer_sandbox.lib.asset_id import AssetId
-from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
-from exasol_script_languages_developer_sandbox.lib.run_setup_ec2 import run_setup_ec2
+from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_setup_ec2 import run_setup_ec2
 
 
 @cli.command()
