+++ exasol_script_languages_developer_sandbox/cli/commands/setup_ec2_and_install_dependencies.py
@@ -8,8 +8,8 @@ from exasol_script_languages_developer_sandbox.cli.options.id_options import id_
 from exasol_script_languages_developer_sandbox.cli.options.logging import logging_options, set_log_level
 from exasol_script_languages_developer_sandbox.lib.ansible.ansible_access import AnsibleAccess
 from exasol_script_languages_developer_sandbox.lib.asset_id import AssetId
-from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
-from exasol_script_languages_developer_sandbox.lib.run_setup_ec2_and_install_dependencies import \
+from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_setup_ec2_and_install_dependencies import \
     run_setup_ec2_and_install_dependencies
