+++ exasol_script_languages_developer_sandbox/cli/commands/setup_vm_bucket.py
@@ -2,8 +2,8 @@ from exasol_script_languages_developer_sandbox.cli.cli import cli
 from exasol_script_languages_developer_sandbox.cli.common import add_options
 from exasol_script_languages_developer_sandbox.cli.options.aws_options import aws_options
 from exasol_script_languages_developer_sandbox.cli.options.logging import logging_options, set_log_level
-from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
-from exasol_script_languages_developer_sandbox.lib.vm_slc_bucket import run_setup_vm_bucket
+from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
+from exasol_script_languages_developer_sandbox.lib.export_vm.vm_slc_bucket import run_setup_vm_bucket
 
 
 @cli.command()
