+++ exasol_script_languages_developer_sandbox/lib/export_vm/run_create_vm.py
@@ -8,13 +8,13 @@ from exasol_script_languages_developer_sandbox.lib.ansible.ansible_repository im
 from exasol_script_languages_developer_sandbox.lib.ansible.ansible_run_context import \
     reset_password_ansible_run_context, default_ansible_run_context
 from exasol_script_languages_developer_sandbox.lib.asset_id import AssetId
-from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
+from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
 
-from exasol_script_languages_developer_sandbox.lib.host_info import HostInfo
-from exasol_script_languages_developer_sandbox.lib.run_export_vm import export_vm
-from exasol_script_languages_developer_sandbox.lib.run_install_dependencies import run_install_dependencies
-from exasol_script_languages_developer_sandbox.lib.run_reset_password import run_reset_password
-from exasol_script_languages_developer_sandbox.lib.run_setup_ec2 import run_lifecycle_for_ec2
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.host_info import HostInfo
+from exasol_script_languages_developer_sandbox.lib.export_vm.run_export_vm import export_vm
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_install_dependencies import run_install_dependencies
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_reset_password import run_reset_password
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_setup_ec2 import run_lifecycle_for_ec2
 
 
 def run_create_vm(aws_access: AwsAccess, ec2_key_file: Optional[str], ec2_key_name: Optional[str],
