+++ exasol_script_languages_developer_sandbox/lib/export_vm/vm_disk_image_format.py
