+++ exasol_script_languages_developer_sandbox/lib/export_vm/vm_slc_bucket.py
@@ -1,6 +1,6 @@
 import logging
 
-from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
+from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
 from exasol_script_languages_developer_sandbox.lib.render_template import render_template
 
 STACK_NAME = "VM-SLC-Bucket"
