+++ exasol_script_languages_developer_sandbox/lib/setup_ec2/host_info.py
