+++ exasol_script_languages_developer_sandbox/lib/setup_ec2/random_string_generator.py
