--- exasol_script_languages_developer_sandbox/templates/aws_tag_value.jinja
@@ -1,1 +1,0 @@
-exa-slc-{{asset_id}}
