--- exasol_script_languages_developer_sandbox/templates/vm_image_name.jinja
@@ -1,1 +1,0 @@
-Exasol-SLC-Developer-Sandbox-{{asset_id}}
