+++ setup.py
@@ -9,6 +9,9 @@ packages = \
  'exasol_script_languages_developer_sandbox.lib',
  'exasol_script_languages_developer_sandbox.lib.ansible',
  'exasol_script_languages_developer_sandbox.lib.asset_printing',
+ 'exasol_script_languages_developer_sandbox.lib.aws_access',
+ 'exasol_script_languages_developer_sandbox.lib.export_vm',
+ 'exasol_script_languages_developer_sandbox.lib.setup_ec2',
  'exasol_script_languages_developer_sandbox.runtime']
 
 package_data = \
