+++ test/test_ansible.py
@@ -8,8 +8,8 @@ from exasol_script_languages_developer_sandbox.lib.ansible.ansible_repository im
     AnsibleResourceRepository
 from exasol_script_languages_developer_sandbox.lib.ansible.ansible_run_context import AnsibleRunContext, \
     default_ansible_run_context
-from exasol_script_languages_developer_sandbox.lib.host_info import HostInfo
-from exasol_script_languages_developer_sandbox.lib.run_install_dependencies import run_install_dependencies
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.host_info import HostInfo
+from exasol_script_languages_developer_sandbox.lib.setup_ec2.run_install_dependencies import run_install_dependencies
 
 import test.ansible
 import test.ansible_conflict
